% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi0_lam.R
\name{pi0est_mc}
\alias{pi0est_mc}
\title{Estimate the proportion of true null hypotheses using Storey's estimation}
\usage{
pi0est_mc(p, d, add, lambda = NULL, lam_sig = 0.1)
}
\arguments{
\item{p}{Vector of p-values for the observed scores. A low p-value corresponds to more evidence against the null.}

\item{d}{The number of decoys.}

\item{add}{Either 0 or 1. Determines what is added to the numerator in the calculation. 0 corresponds to Storey's asymptotic formulation, while 1 corresponds to the finite correction.}

\item{lambda}{The value of the tuning parameter in the estimate. Either a number between 0 and 1 (non-inclusive) or \code{NULL} with the latter corresponding to using
the binomial test method to select \eqn{\lambda} according to the data.}

\item{lam_sig}{The significance level of the binomial test to choose \eqn{\lambda}.}
}
\value{
\item{pi_0_est}{The estimated value of \eqn{\pi_0}.}
\item{lambda}{The value of lambda used to calculated \eqn{\pi_0}.}
}
\description{
Estimates the proportion of true nulls in a vector of p-values using the estimation provided by Storey (2002) and Storey, Taylor & Seigmund (2004). Uses either a pre-set value
of \eqn{\lambda} or selects one based off the data.
Can be set to use a correction for finite hypotheses.
}
\references{
Storey JD. (2002) A direct approach to false discovery rates. Journal
of the Royal Statistical Society, Series B, 64: 479-498. \cr
\url{http://onlinelibrary.wiley.com/doi/10.1111/1467-9868.00346/abstract}

Storey JD, Taylor JE, and Siegmund D. (2004) Strong control,
conservative point estimation, and simultaneous conservative
consistency of false discovery rates: A unified approach. Journal of
the Royal Statistical Society, Series B, 66: 187-205. \cr
\url{http://onlinelibrary.wiley.com/doi/10.1111/j.1467-9868.2004.00439.x/abstract}

Emery K, Keich U, Hasam S and Nobel W. (2019) Multiple competition based FDR control. arXiv:1907.01458\cr
\url{https://arxiv.org/abs/1907.01458}
}
\author{
Kristen Emery
}
