\name{coef.cor}
\alias{coef.cor}
\alias{conc2case}
\alias{bicomprisk}
\alias{test.conc}
\alias{back2timereg}
\alias{or.cif}
\alias{print.cor}
\alias{summary.randomcif}
\alias{print.summary.cor}
\alias{coef.randomcif}
\alias{print.randomcif}
\alias{summary.randomcifrv}
\alias{coef.randomcifrv}
\alias{print.randomcifrv}
\alias{print.testconc}
\alias{multcif}
\alias{plack.cif}
\alias{plack.cif2}
\alias{Specials}
\title{For internal use}
\description{for internal use}
\author{Thomas Scheike}
\keyword{survival}
