\name{coef.cor}
\alias{coef.cor}
\alias{summary.cor}
\alias{print.cor}
\alias{summary.randomcif}
\alias{coef.randomcif}
\alias{print.randomcif}
\alias{summary.randomcifrv}
\alias{coef.randomcifrv}
\alias{print.randomcifrv}
\alias{multcif}
\alias{plack.cif}
\alias{plack.cif2}
\title{For internal use}
\description{for internal use}
\author{Thomas Scheike}
\keyword{survival}
